despachar

+++ app/Infrastructure/Persistence/Repository/OrdenProduccionRepository.php
@@ -2,16 +2,22 @@
 
 namespace App\Infrastructure\Persistence\Repository;
 
+use App\Infrastructure\Persistence\Model\ProduccionBatch as ProduccionBatchModel;
 use App\Infrastructure\Persistence\Model\OrdenProduccion as OrdenProduccionModel;
 use App\Domain\Produccion\Aggregate\OrdenProduccion as AggregateOrdenProduccion;
 use App\Domain\Produccion\Aggregate\ProduccionBatch as AggregateProduccionBatch;
 use App\Infrastructure\Persistence\Model\OrderItem as OrdenProduccionItemModel;
+use App\Infrastructure\Persistence\Model\ItemDespacho as ItemDespachoModel;
+use App\Domain\Produccion\Aggregate\ItemDespacho as AggregateItemDespacho;
+use App\Infrastructure\Persistence\Repository\ProduccionBatchRepository;
 use App\Domain\Produccion\Repository\OrdenProduccionRepositoryInterface;
+use App\Infrastructure\Persistence\Repository\ItemDespachoRepository;
 use App\Domain\Produccion\Aggregate\OrdenItem as AggregateOrdenItem;
 use App\Infrastructure\Persistence\Repository\OrdenItemRepository;
 use App\Domain\Produccion\Model\OrderItems as ModelOrderItems;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use App\Domain\Produccion\Aggregate\EstadoPlanificado;
+use App\Domain\Produccion\ValueObjects\ItemDespacho;
 use App\Domain\Produccion\ValueObjects\OrderItem;
 use App\Domain\Produccion\Aggregate\EstadoOP;
 use App\Domain\Produccion\ValueObjects\Qty;
@@ -27,12 +33,30 @@ class OrdenProduccionRepository implements OrdenProduccionRepositoryInterface
     public readonly OrdenItemRepository $ordenItemRepository;
 
     /**
+     * @var ItemDespachoRepository
+     */
+    public readonly ItemDespachoRepository $itemDespachoRepository;
+
+    /**
+     * @var ProduccionBatchRepository
+     */
+    public readonly ProduccionBatchRepository $produccionBatchRepository;
+
+    /**
      * Constructor
      * 
      * @param OrdenItemRepository $ordenItemRepository
+     * @param ItemDespachoRepository $itemDespachoRepository
+     * @param ProduccionBatchRepository $produccionBatchRepository
      */
-    public function __construct(OrdenItemRepository $ordenItemRepository) {
+    public function __construct(
+        OrdenItemRepository $ordenItemRepository,
+        ItemDespachoRepository $itemDespachoRepository,
+        ProduccionBatchRepository $produccionBatchRepository
+    ) {
         $this->ordenItemRepository = $ordenItemRepository;
+        $this->itemDespachoRepository = $itemDespachoRepository;
+        $this->produccionBatchRepository = $produccionBatchRepository;
     }
 
     /**
@@ -50,8 +74,9 @@ class OrdenProduccionRepository implements OrdenProduccionRepositoryInterface
 
         $fecha = $this->mapDateToDomain($row->fecha);
         $estado = EstadoOP::from($row->estado);
-        $items = $this->mapItemsToDomain($row->items);
+        $items = $this->mapItems($row->items);
         $batches = $this->mapItemsBatches($row->batches);
+        $itemsDespacho = $this->mapItemsDespachos($row->despachoItems);
 
         return AggregateOrdenProduccion::reconstitute(
             $row->id,
@@ -59,7 +84,8 @@ class OrdenProduccionRepository implements OrdenProduccionRepositoryInterface
             $row->sucursal_id,
             $estado,
             $items,
-            $batches
+            $batches,
+            $itemsDespacho
         );
     }
 
@@ -82,8 +108,20 @@ class OrdenProduccionRepository implements OrdenProduccionRepositoryInterface
         $orderId = $model->id;
 
         if ($resetItems) {
-            OrdenProduccionItemModel::query()->where('op_id', $orderId)->delete();
-            $this->mapItemsToRows($orderId, $op->items());
+            if ($op->estado()->value == EstadoOP::CREADA->value) {
+                OrdenProduccionItemModel::query()->where('op_id', $orderId)->delete();
+                $this->savedItems($orderId, $op->items());
+            }
+
+            if ($op->estado()->value == EstadoOP::PLANIFICADA->value) {
+                ProduccionBatchModel::query()->where('op_id', $orderId)->delete();
+                $this->savedBatch($op->batches());
+            }
+
+            if ($op->estado()->value == EstadoOP::CERRADA->value) {
+                ItemDespachoModel::query()->where('op_id', $orderId)->delete();
+                $this->savedDespacho($op->itemsDespacho());
+            }
         }
 
         if ($sendOutbox) {
@@ -93,33 +131,35 @@ class OrdenProduccionRepository implements OrdenProduccionRepositoryInterface
         return $orderId;
     }
 
-    /** 
+    /**
+     * @param mixed $data
      * @return ModelOrderItems
      */
-    private function mapItemsToDomain($items): ModelOrderItems
+    private function mapItems($data): ModelOrderItems
     {
-        $domainItems = [];
+        $items = [];
 
-        foreach ($items as $row) {
-            $domainItems[] = new OrderItem(
-                new Sku($row->product->sku),
+        foreach ($data as $row) {
+            $items[] = new OrderItem(
+                new Sku(value: $row->product->sku),
                 new Qty($row->qty),
                 $row->p_id
             );
         }
 
-        return ModelOrderItems::fromArray($domainItems);
+        return ModelOrderItems::fromArray($items);
     }
 
     /**
+     * @param mixed $data
      * @return AggregateProduccionBatch[]
      */
-    private function mapItemsBatches($batches): array
+    private function mapItemsBatches($data): array
     {
-        $domainItems = [];
+        $items = [];
 
-        foreach ($batches as $row) {
-            $domainItems[] = new AggregateProduccionBatch(
+        foreach ($data as $row) {
+            $items[] = new AggregateProduccionBatch(
                 $row->id,
                 $row->op_id,
                 $row->p_id,
@@ -137,7 +177,26 @@ class OrdenProduccionRepository implements OrdenProduccionRepositoryInterface
             );
         }
 
-        return $domainItems;
+        return $items;
+    }
+
+    /**
+     * @param mixed $data
+     * @return ItemDespacho[]
+     */
+    private function mapItemsDespachos($data): array
+    {
+        $items = [];
+
+        foreach ($data as $row) {
+            $items[] = new ItemDespacho(
+                $row->op_id,
+                $row->product_id,
+                $row->paquete_id
+            );
+        }
+
+        return $items;
     }
 
     /**
@@ -145,7 +204,7 @@ class OrdenProduccionRepository implements OrdenProduccionRepositoryInterface
      * @param ModelOrderItems $items
      * @return void
      */
-    private function mapItemsToRows(int|null $opId, ModelOrderItems $items): void
+    private function savedItems(int|null $opId, ModelOrderItems $items): void
     {
         foreach ($items as $item) {
             $this->ordenItemRepository->save(
@@ -157,6 +216,51 @@ class OrdenProduccionRepository implements OrdenProduccionRepositoryInterface
                     $item->qty()->value(),
                     0,
                     0
+                )
+            );
+        }
+    }
+
+    /**
+     * @param int|null $opId
+     * @param array $items
+     * @return void
+     */
+    private function savedBatch(array $items): void
+    {
+        foreach ($items as $key => $item) {
+            $this->produccionBatchRepository->save(
+                new AggregateProduccionBatch(
+                    null,
+                    $item->ordenProduccionId,
+                    $item->productoId,
+                    $item->estacionId,
+                    $item->recetaVersionId,
+                    $item->porcionId,
+                    $item->cantPlanificada,
+                    $item->cantProducida,
+                    $item->mermaGr,
+                    $item->estado,
+                    $item->rendimiento,
+                    $item->qty,
+                    $key + 1
+                )
+            );
+        }
+    }
+
+    /**
+     * @param array $items
+     * @return void
+     */
+    private function savedDespacho(array $items): void
+    {
+        foreach ($items as $item) {
+            $this->itemDespachoRepository->save(
+                new AggregateItemDespacho(
+                    $item->ordenProduccionId,
+                    $item->productId,
+                    null
                 )
             );
         }
